+++ src/main/java/lotto/domain/purchase/InvalidPurchasingCountException.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoMoney;
+package lotto.domain.purchase;
 
 public class InvalidPurchasingCountException extends IllegalArgumentException {
 	public static final String INVALID = "구매 가능한 숫자가 아닙니다.";
