+++ src/main/java/lotto/domain/purchase/LottoMoney.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoMoney;
+package lotto.domain.purchase;
 
 import java.util.Objects;
 
@@ -12,6 +12,21 @@ public class LottoMoney {
 	public LottoMoney(long lottoMoney) {
 		validate(lottoMoney);
 		this.lottoMoney = lottoMoney;
+	}
+
+	public static LottoMoney valueOf(String inputLottoMoney) {
+		try {
+			validateNullOrEmpty(inputLottoMoney);
+			return new LottoMoney(Long.parseLong(inputLottoMoney));
+		} catch (NumberFormatException e) {
+			throw new InvalidLottoMoneyException(InvalidLottoMoneyException.NOT_INTEGER);
+		}
+	}
+
+	private static void validateNullOrEmpty(String inputLottoMoney) {
+		if (Objects.isNull(inputLottoMoney) || inputLottoMoney.isEmpty()) {
+			throw new InvalidLottoMoneyException(InvalidLottoMoneyException.NULL_OR_EMPTY);
+		}
 	}
 
 	private void validate(long lottoMoney) {
@@ -33,21 +48,6 @@ public class LottoMoney {
 
 	private boolean isDivideByUnit(long lottoMoney) {
 		return (lottoMoney % UNIT) != 0L;
-	}
-
-	public static LottoMoney valueOf(String inputLottoMoney) {
-		try {
-			validateNullOrEmpty(inputLottoMoney);
-			return new LottoMoney(Long.parseLong(inputLottoMoney));
-		} catch (NumberFormatException e) {
-			throw new InvalidLottoMoneyException(InvalidLottoMoneyException.NOT_INTEGER);
-		}
-	}
-
-	private static void validateNullOrEmpty(String inputLottoMoney) {
-		if (Objects.isNull(inputLottoMoney) || inputLottoMoney.isEmpty()) {
-			throw new InvalidLottoMoneyException(InvalidLottoMoneyException.NULL_OR_EMPTY);
-		}
 	}
 
 	public PurchasingCount generatePurchasingLottoTicketCount() {
