+++ src/main/java/lotto/domain/result/InvalidMatchCountException.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoRank;
+package lotto.domain.result;
 
 public class InvalidMatchCountException extends IllegalArgumentException {
 	public static final String OUT_OF_BOUND = "로또 번호의 일치하는 개수가 유효하지 않습니다.";
