+++ src/main/java/lotto/domain/result/MatchCount.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoRank;
+package lotto.domain.result;
 
 import java.util.Objects;
