+++ src/test/java/lotto/domain/lottoTicket/AutoLottoTicketsFactoryTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import lotto.domain.lottoMoney.PurchasingCount;
+import lotto.domain.purchase.PurchasingCount;
 
 class AutoLottoTicketsFactoryTest {
 	@ParameterizedTest
