+++ src/test/java/lotto/domain/purchase/ManualLottoTicketCountTest.java
@@ -1,12 +1,10 @@
-package lotto.domain.lottoTicket;
+package lotto.domain.purchase;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import lotto.domain.lottoMoney.PurchasingCount;
 
 class ManualLottoTicketCountTest {
 	@ParameterizedTest
