+++ src/test/java/lotto/domain/result/MatchCountTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoRank;
+package lotto.domain.result;
 
 import static org.assertj.core.api.Assertions.*;
